computer-vision-advanced/facial-verification/api/pyproject.toml
[project]
name = "facial-verification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "chromadb",
    "pymilvus",
    "insightface",
    "onnxruntime-gpu",
    "opencv-python",
    "pillow",
    "fastapi[standard]",
    "python-multipart",
    "websocket",
]
